squid-proxy/tests/requirements.txt
requests
python-dotenv
curl_cffi
